+++ src/kimi_cli/metadata.py
@@ -6,6 +6,7 @@ from pathlib import Path
 
 from pydantic import BaseModel, Field
 
+from kaos.path import KaosPath
 from kimi_cli.share import get_share_dir
 from kimi_cli.utils.logging import logger
 
@@ -25,9 +26,12 @@ class WorkDirMeta(BaseModel):
 
     @property
     def sessions_dir(self) -> Path:
-        path = get_share_dir() / "sessions" / md5(self.path.encode(encoding="utf-8")).hexdigest()
-        path.mkdir(parents=True, exist_ok=True)
-        return path
+        """The directory to store sessions for this work directory."""
+        basename = KaosPath(self.path).name
+        path_md5 = md5(self.path.encode(encoding="utf-8")).hexdigest()
+        session_dir = get_share_dir() / "sessions" / f"{basename}_{path_md5}"
+        session_dir.mkdir(parents=True, exist_ok=True)
+        return session_dir
 
 
 class Metadata(BaseModel):
@@ -38,6 +42,19 @@ class Metadata(BaseModel):
 
     thinking: bool = False
     """Whether the last session was in thinking mode."""
+
+    def get_work_dir_meta(self, path: KaosPath) -> WorkDirMeta | None:
+        """Get the metadata for a work directory."""
+        for wd in self.work_dirs:
+            if wd.path == str(path):
+                return wd
+        return None
+
+    def new_work_dir_meta(self, path: KaosPath) -> WorkDirMeta:
+        """Create a new work directory metadata."""
+        wd_meta = WorkDirMeta(path=str(path))
+        self.work_dirs.append(wd_meta)
+        return wd_meta
 
 
 def load_metadata() -> Metadata: